+++ src/Core/src/Platform/Android/GlobalWindowInsetListener.cs
@@ -167,92 +167,10 @@ namespace Microsoft.Maui.Platform
     }
 }
 
-#pragma warning disable RS0016
-public static class SafeAreaExtension
-{
-    internal static ISafeAreaView2? GetSafeAreaView2(object? layout) =>
-                layout switch
-                {
-                    ISafeAreaView2 sav2 => sav2,
-                    IElementHandler { VirtualView: ISafeAreaView2 virtualSav2 } => virtualSav2,
-                    _ => null
-                };
-
-    internal static ISafeAreaView? GetSafeAreaView(object? layout) =>
-        layout switch
-        {
-            ISafeAreaView sav => sav,
-            IElementHandler { VirtualView: ISafeAreaView virtualSav } => virtualSav,
-            _ => null
-        };
-
-
-    internal static SafeAreaRegions GetSafeAreaRegionForEdge(int edge, ICrossPlatformLayout crossPlatformLayout)
-    {
-        var layout = crossPlatformLayout;
-        var safeAreaView2 = GetSafeAreaView2(layout);
-
-        if (safeAreaView2 is not null)
-        {
-            return safeAreaView2.GetSafeAreaRegionsForEdge(edge);
-        }
-
-        var safeAreaView = GetSafeAreaView(layout);
-        return safeAreaView?.IgnoreSafeArea == false ? SafeAreaRegions.Container : SafeAreaRegions.None;
-    }
-
-    internal static SafeAreaPadding GetAdjustedSafeAreaInsets(WindowInsetsCompat windowInsets, ICrossPlatformLayout crossPlatformLayout, Context context)
-    {
-        var baseSafeArea = windowInsets.ToSafeAreaInsets(context);
-        var keyboardInsets = windowInsets.GetKeyboardInsets(context);
-        var isKeyboardShowing = !keyboardInsets.IsEmpty;
-
-        var layout = crossPlatformLayout;
-        var safeAreaView2 = GetSafeAreaView2(layout);
-
-        if (safeAreaView2 is not null)
-        {
-            // Apply safe area selectively per edge based on SafeAreaRegions
-            var left = GetSafeAreaForEdge(GetSafeAreaRegionForEdge(0, layout), baseSafeArea.Left, 0, isKeyboardShowing, keyboardInsets);
-            var top = GetSafeAreaForEdge(GetSafeAreaRegionForEdge(1, layout), baseSafeArea.Top, 1, isKeyboardShowing, keyboardInsets);
-            var right = GetSafeAreaForEdge(GetSafeAreaRegionForEdge(2, layout), baseSafeArea.Right, 2, isKeyboardShowing, keyboardInsets);
-            var bottom = GetSafeAreaForEdge(GetSafeAreaRegionForEdge(3, layout), baseSafeArea.Bottom, 3, isKeyboardShowing, keyboardInsets);
-
-            return new SafeAreaPadding(left, right, top, bottom);
-        }
-
-        // Fallback: return the base safe area for legacy views
-        return baseSafeArea;
-    }
-
-    internal static double GetSafeAreaForEdge(SafeAreaRegions safeAreaRegion, double originalSafeArea, int edge, bool isKeyboardShowing, SafeAreaPadding keyBoardInsets)
-    {
-        // Edge-to-edge content - no safe area padding
-        if (safeAreaRegion == SafeAreaRegions.None)
-        {
-            return 0;
-        }
-
-        // Handle SoftInput specifically - only apply keyboard insets for bottom edge when keyboard is showing
-        if (SafeAreaEdges.IsSoftInput(safeAreaRegion) && isKeyboardShowing && edge == 3)
-        {
-            return keyBoardInsets.Bottom;
-        }
-
-        // All other regions respect safe area in some form
-        // This includes:
-        // - Default: Platform default behavior
-        // - All: Obey all safe area insets  
-        // - Container: Content flows under keyboard but stays out of bars/notch
-        // - Any combination of the above flags
-        return originalSafeArea;
-    }
-}
-
 /// <summary>
 /// Extension methods to access the shared GlobalWindowInsetListener instance.
 /// </summary>
-public static class GlobalWindowInsetListenerExtensions
+internal static class GlobalWindowInsetListenerExtensions
 {
     /// <summary>
     /// Gets the shared GlobalWindowInsetListener instance from the current MauiAppCompatActivity.
